server/requirements.txt
Flask>=3.1.1
flask_cors>=6.0.1
numpy>=2.3.1
PyYAML>=6.0.2
rdflib>=7.1.4
Requests>=2.32.4
scikit_learn>=1.7.1
textdistance>=4.6.3
tiktoken>=0.9.0
